Move functions from run_util.go to better places
